Refactor JugadorVolley model and update controller methods to use Long for ID; adjust application properties for Hibernate configuration

+++ src/main/java/com/example/demo/Repository/IJugadorVolleyRepository.java
@@ -7,6 +7,6 @@ import com.example.demo.Model.JugadorVolley;
 
 
 @Repository
-public interface IJugadorVolleyRepository extends JpaRepository<JugadorVolley, Integer> {
+public interface IJugadorVolleyRepository extends JpaRepository<JugadorVolley, Long> {
 
 }
